v0.0.04 - Input and Scoring
Reintroduced scoring, and matched inputs to the buttons.

Created InputController, that monitors the buttons being pressed. Tried to create a global input controller (KeyboardInputController) based on Yu Chao's files, but failed, so I implemented a basic one for now while the game gets fleshed out. It will need to be overhauled as the keyboard inputs are hardcoded right now

Added colliders to the note prefab, and programmed InputController to listen to inputs and send them to the buttons.

Arrows now have colliders and recognized when they hit one of the buttons. button hit recognition is controlled through NoteController. Not sure if this is how it should be, but I put it in there anyways for now.

Added a function in NoteController to initialize the rotation. There's gotta be an easier way to do this. Maybe dictionaries? Track information is now also passed from BeatScroller

Need to start calling BeatScroller the conductor.

Need to add masking for the notess that dynamically changes based on scoring

The beat seems slightly off. I need to play with how notes are spwaned and the beats they're spawned on.

Need to add fills and solos

will probably run into issues as notes get closer with overlapping notes and double hits. Hopefully that never happens