v0.0.06
Updated aspect ratio so builds actually render to correct screen sizes

Moved buttons to still be in the window in regular screens

modified NoteController to look at the distance to the FinishLineX in BeatScroller instead of hard code.

Need to adjust this to beat instead of distance. Should remove collider

Added overlay sprite to mask note spaawns. Will change it later. For now, you're fighting a fire-extinguisher.