v0.0.10 - Editor!
Added the editor scene. Allows you to modify songs in their SongInfo object using a GUI instead of in notepad.

Scroll to change the beat
Hold Shift to switch to fills

Need to addd checks so that fills and notes don't occur on the same beat.

Honestly I'm thrilled with this. I didn't think I could do it.