v0.0.11 - Editor advancement, menu, UI
Worked a lot on the editor, making it functional, and adding the ability to play the song from wherever you are. There's still problems switching to playing with the song time and the way beats arre spaced out. Need to spend more time on how dsptime works.

Added a main menu, and the ability to move back and forth between different scenes in the game. This is really just for making a build to show Colin for now, but is a good lesson in scene management. Later it'll be important to play around with scenes. Still need to learn how long-term variables work.

Did an update that fixes the timing and beats on the editor, and reworked the main conductor to load the edited song. Learning a ton by doing this and it's awesome.